refactor(core): create core crate with shared types
