+++ crates/acrawl-cli/Cargo.toml
@@ -10,6 +10,7 @@ name = "acrawl"
 path = "src/main.rs"
 
 [dependencies]
+acrawl-core = { path = "../core" }
 api = { path = "../api" }
 commands = { path = "../commands" }
 crawler = { path = "../crawler" }
