+++ crates/api/Cargo.toml
@@ -6,6 +6,7 @@ license.workspace = true
 publish.workspace = true
 
 [dependencies]
+acrawl-core = { path = "../core" }
 aws-credential-types = { version = "1", features = ["hardcoded-credentials"] }
 aws-sigv4 = { version = "1", features = ["sign-http"] }
 azure_identity = { version = "0.34", optional = true }
