+++ crates/core/Cargo.toml
@@ -1,13 +1,13 @@
 [package]
-name = "commands"
+name = "acrawl-core"
 version.workspace = true
 edition.workspace = true
 license.workspace = true
 publish.workspace = true
 
 [dependencies]
-acrawl-core = { path = "../core" }
-runtime = { path = "../runtime" }
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
 
 [lints]
 workspace = true
