+++ crates/core/src/lib.rs
@@ -1,0 +1,7 @@
+pub mod event;
+pub mod message;
+pub mod tool_spec;
+
+pub use event::AssistantEvent;
+pub use message::{ContentBlock, ConversationMessage, MessageRole, TokenUsage};
+pub use tool_spec::ToolSpec;
