+++ crates/runtime/Cargo.toml
@@ -6,6 +6,7 @@ license.workspace = true
 publish.workspace = true
 
 [dependencies]
+acrawl-core = { path = "../core" }
 getrandom = "0.2"
 sha2 = "0.10"
 reqwest = { version = "0.12", features = ["rustls-tls", "json"], default-features = false }
